🔨 Refactor(#163): 전역상태 stores로 수정
