+++ src/features/quiz/ui/MultipleChoice.tsx
@@ -4,7 +4,7 @@ import {
   MultipleChoiceButtonDiv,
   Img,
 } from './styles';
-import { useClientQuizStore } from '@/features/quiz/useClientQuizStore';
+import { useClientQuizStore } from '@/features/quiz/stores';
 import type { Quiz } from '@features/quiz/types';
 import { useElementRect } from '@/features/intro/service/hooks';
 import { getImageUrl } from '@/utils/getImageUrl';
